+++ src/Components/Alert/PopUp/UploadProject/index.tsx
@@ -1,20 +1,20 @@
 import React, { memo, useCallback, useMemo, useState } from 'react';
 
-import { CredentialDTO, Loading } from '@Types/AppTypes';
+import { CredentialDTO } from '@Types/AppTypes';
 import { ProjectDTO } from '@Types/ProjectTypes';
 import { navigate } from '@Globals/NavigationControler';
+import { translations } from '@Translations/index';
 import ConfigService from '@Services/ConfigService';
-import DateTimeService from '@Services/DateTimeService';
 import ProjectService from '@Services/ProjectService';
 import AlertService from '@Services/AlertService';
 import CacheService from '@Services/CacheService';
 import SyncService from '@Services/SyncService';
-import AppAPI from '@appAPI/index';
+import FetchAPIService from '@APIServices/FetchAPIService';
+import DataProcessService from '@APIServices/DataProcessService';
 
 import { LC } from '@Alert/__LC__';
 import { CredentialsDisplay } from './CredentialsDisplay';
 import { ErrorDisplay } from './ErrorDisplay';
-import { LoadingDisplay } from './LoadingDisplay';
 import { FooterButtons } from './FooterButtons';
 
 export const UploadProjects = memo((props: {
@@ -22,14 +22,17 @@ export const UploadProjects = memo((props: {
   closeModal: () => void
 }) => {
 
-  const config     = useMemo(() => ConfigService.config, []);
   const controller = useMemo(() => new AbortController(), []);
-  const [error   , setError  ] = useState<string | null>(null);
-  const [loading , setLoading] = useState<Loading>('Loaded');
-
-  const showCredentialsDisplay = error === null && loading === 'Loaded';
-  const showErrorDisplay       = error !== null && loading === 'Loaded';
-  const showLoadingDisplay     = loading === 'Loading';
+  const config     = useMemo(() => ConfigService.config, []);
+  const RS         = useMemo(() => translations.component.alert.shared[config.language], []);
+  const R          = useMemo(() => translations.component.alert.uploadProject[config.language], []);
+  const [error    , setError    ] = useState<string | null>(null);
+  const [feedbacks, setFeedbacks] = useState<string[]>([]);
+  const [show     , setShow     ] = useState({
+    credentialsDisplay: true,
+    loadingDisplay: false,
+    errorDisplay: false,
+  });
 
   const onCancel = useCallback(() => {
     props.closeModal();
@@ -37,46 +40,45 @@ export const UploadProjects = memo((props: {
   }, [props.closeModal]);
 
   const onCredentialChoose = useCallback(async (credential: CredentialDTO) => {
-    setLoading('Loading');
-    const project = await ProjectService.buildProjectFromDatabase(props.id_project);
-    await uploadProject(credential, project);
-  }, [props.id_project]);
 
-  const uploadProject = useCallback(async (credential: CredentialDTO, projectDTO: ProjectDTO) => {
-    const processedProject = dataProcessingBeforeUpload(credential, projectDTO);
-    await new AppAPI(credential).uploadProject(controller.signal, processedProject,
+    setFeedbacks([]);
+    setShow(prev => ({ ...prev,
+      loadingDisplay: true,
+      credentialsDisplay: false,
+    }));
+
+    const projectDTO = await DataProcessService.buildProjectFromDatabase(props.id_project,
+      (feedbackMessage) => setFeedbacks(prev => ([ ...prev, feedbackMessage ]))
+    );
+
+    const processedProject = DataProcessService.attachUploadEntry(credential, projectDTO, {
+      dateFormat: config.dateFormat,
+      timeFormat: config.timeFormat,
+    }, (feedbackMessage) => setFeedbacks(prev => ([ ...prev, feedbackMessage ])));
+
+    await new FetchAPIService(credential).uploadProject(controller.signal, processedProject,
+      (feedbackMessage) => setFeedbacks(prev => ([ ...prev, feedbackMessage ])),
       async () => {
-        await dataProcessingAfterUpload(processedProject);
+        await afterUploadProcessing(processedProject);
+        setFeedbacks(prev => [ ...prev, RS['Done!']]);
         await AlertService.runAcceptCallback();
         props.closeModal();
       },
       (errorMessage) => {
         setError(errorMessage);
-        setLoading('Loaded');
+        setFeedbacks(prev => [ ...prev, RS['Error!']]);
+        setShow(prev => ({ ...prev, errorDisplay: true }));
       }
     );
-  }, []);
 
-  const dataProcessingBeforeUpload = useCallback((credential: CredentialDTO, projectDTO: ProjectDTO) => {
+  }, [props.id_project]);
 
-    // Upload Date/Time entry ================
-    projectDTO.projectSettings.uploads ??= [];
-    projectDTO.projectSettings.uploads.push({
-      dateUTM: DateTimeService.getCurrentDateTimeUTC(),
-      date:    DateTimeService.getCurrentDateTime(config),
-      url:     credential.rootURL,
-    });
-
-    return projectDTO;
-
-  }, []);
-
-  const dataProcessingAfterUpload = useCallback(async (projectDTO: ProjectDTO) => {
+  const afterUploadProcessing = useCallback(async (projectDTO: ProjectDTO) => {
 
     const { rules, id_project } = projectDTO.projectSettings;
 
-    // Project deletion rule ==============
     if (rules.deleteAfterUpload === true) {
+      setFeedbacks(prev => [ ...prev, R['Deleting project']]);
       await ProjectService.deleteProject(id_project,
         async () => {
           if (CacheService.lastOpenProject.id_project === id_project) {
@@ -86,29 +88,33 @@ export const UploadProjects = memo((props: {
         },
         (errorMessage) => {
           setError(errorMessage);
-          setLoading('Loaded');
+          setFeedbacks(prev => [ ...prev, RS['Error!']]);
+          setShow(prev => ({ ...prev, errorDisplay: true }));
         }
       );
       return;
     }
 
-    SyncService.updateSyncDataAfterUpload(id_project);
-    setLoading('Loaded');
+    if (!rules.deleteAfterUpload) {
+      setFeedbacks(prev => [ ...prev, R['Updating sync data file']]);
+      SyncService.updateSyncDataAfterUpload(id_project);
+    }
 
   }, []);
 
   return (
     <LC.PopUp>
       <CredentialsDisplay
-        showDisplay={showCredentialsDisplay}
+        showDisplay={show.credentialsDisplay}
         onCredentialChoose={async (credential) => await onCredentialChoose(credential)}
       />
       <ErrorDisplay
-        showDisplay={showErrorDisplay}
+        showDisplay={show.errorDisplay}
         error={error}
       />
-      <LoadingDisplay
-        showDisplay={showLoadingDisplay}
+      <LC.Feedback
+        showDisplay={show.loadingDisplay}
+        feedbackMessage={feedbacks}
       />
       <FooterButtons
         onCancel={() => onCancel()}