+++ src/Components/Alert/__LC__/index.tsx
@@ -1,5 +1,7 @@
 import { PopUp } from './PopUp';
+import { Feedback } from './Feedback';
 
 export const LC = {
   PopUp: PopUp,
+  Feedback: Feedback,
 };