+++ src/Translations/component/Alert/DownloadProject.ts
@@ -2,22 +2,19 @@ import { LanguageTag } from '@Types/AppTypes';
 
 type TranslationDTO = Record<LanguageTag, {
   'Download new projects from?': string
-  'Error': string
-  'Connecting...': string
   'Projects available:': string
+  'Starting processing': string
 }>
 
 export const R_Alert_DownloadProjects: TranslationDTO = {
   'en-US': {
     'Download new projects from?': 'Download new projects from?',
-    'Error': 'Error',
-    'Connecting...': 'Connecting...',
     'Projects available:': 'Projects available:',
+    'Starting processing': 'Starting processing',
   },
   'pt-BR': {
     'Download new projects from?': 'Escolha onde fazer donwload?',
-    'Error': 'Erro',
-    'Connecting...': 'Conectando...',
     'Projects available:': 'Projetos disponíveis:',
+    'Starting processing': 'Inciando processamento',
   },
 };