+++ src/Translations/component/Alert/UploadProject.ts
@@ -2,19 +2,19 @@ import { LanguageTag } from '@Types/AppTypes';
 
 type TranslationDTO = Record<LanguageTag, {
   'Upload this project to?': string
-  'Error': string
-  'Connecting...': string
+  'Deleting project': string
+  'Updating sync data file': string
 }>
 
 export const R_Alert_UploadProject: TranslationDTO = {
   'en-US': {
     'Upload this project to?': 'Upload this project to?',
-    'Error': 'Error',
-    'Connecting...': 'Connecting...',
+    'Deleting project': 'Deleting project',
+    'Updating sync data file': 'Updating sync data file',
   },
   'pt-BR': {
     'Upload this project to?': 'Enviar este projeto para?',
-    'Error': 'Erro',
-    'Connecting...': 'Conectando...',
+    'Deleting project': 'Deletando projeto',
+    'Updating sync data file': 'Atualizando arquivo de syncronização',
   },
 };